ino and matlab files
The ino is for ESP8266, to wirelessly tranmit data. This works, just that the sync cant be establish. we will use it later for telemetry data. rate of 1-5-10 msec was desired but packets are getting lost.
higher rate will definatly work

there are 2 matlab files . one for UDP recive.\, other for serial recive. i will use the serial recive for control development for now, but the UDP one work well with ESP8266, with the constaraint stated above